requirements.txt
anthropic>=0.40.0
faster-whisper>=1.0.0
edge-tts>=6.1.0
sounddevice>=0.5.0
soundfile>=0.12.0
keyboard>=0.13.5
httpx>=0.27.0
python-dotenv>=1.0.0
rich>=13.7.0
